rename package name from CommandImpl to impl
