+++ dubbo-plugin/dubbo-qos/src/main/java/com/alibaba/dubbo/qos/command/impl/Quit.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alibaba.dubbo.qos.command.CommandImpl;
+package com.alibaba.dubbo.qos.command.impl;
 
 import com.alibaba.dubbo.qos.command.BaseCommand;
 import com.alibaba.dubbo.qos.command.CommandContext;
