+++ dubbo-plugin/dubbo-qos/src/main/resources/META-INF/dubbo/internal/com.alibaba.dubbo.qos.command.BaseCommand
@@ -1,5 +1,5 @@
-online=com.alibaba.dubbo.qos.command.CommandImpl.Online
-help=com.alibaba.dubbo.qos.command.CommandImpl.Help
-quit=com.alibaba.dubbo.qos.command.CommandImpl.Quit
-ls=com.alibaba.dubbo.qos.command.CommandImpl.Ls
-offline=com.alibaba.dubbo.qos.command.CommandImpl.Offline
+online=com.alibaba.dubbo.qos.command.impl.Online
+help=com.alibaba.dubbo.qos.command.impl.Help
+quit=com.alibaba.dubbo.qos.command.impl.Quit
+ls=com.alibaba.dubbo.qos.command.impl.Ls
+offline=com.alibaba.dubbo.qos.command.impl.Offline
